Move uniforms-material LongTextField test to suite
